Split tests out of VerifyClientTest for improved readability.
