+++ src/test/java/com/nexmo/client/verify/VerifyClientTest.java
@@ -71,126 +71,6 @@ public class VerifyClientTest {
         return result;
     }
 
-    @Test
-    public void testVerify() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(200, "{\n" +
-                "  \"request_id\": \"not-really-a-request-id\",\n" +
-                "  \"status\": 0,\n" +
-                "  \"error_text\": \"error\"\n" +
-                "}"));
-        VerifyResult r = client.verify(
-                "447700900999",
-                "TestBrand",
-                "15555215554", 6, Locale.US);
-        assertEquals(VerifyResult.STATUS_OK, r.getStatus());
-    }
-
-    @Test
-    public void testVerifyWithRequestObject() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(200, "{\n" +
-                "  \"request_id\": \"not-really-a-request-id\",\n" +
-                "  \"status\": 0,\n" +
-                "  \"error_text\": \"error\"\n" +
-                "}"));
-        VerifyResult r = client.verify(new VerifyRequest(
-                "447700900999",
-                "TestBrand",
-                "15555215554", 6, Locale.US));
-        assertEquals(VerifyResult.STATUS_OK, r.getStatus());
-    }
-
-    @Test
-    public void testVerifyHttpError() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(500, "{\n" +
-                "  \"request_id\": \"not-really-a-request-id\",\n" +
-                "  \"status\": 0,\n" +
-                "  \"error_text\": \"error\"\n" +
-                "}"));
-        try {
-            client.verify(
-                    "447700900999",
-                    "TestBrand",
-                    "15555215554", 6, Locale.US);
-            fail("An IOException should be thrown if an HTTP 500 response is received.");
-        } catch (IOException ioe) {
-            // This is expected
-        }
-    }
-
-    @Test
-    public void testVerifyWithNonNumericStatus() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(200, "{\n" +
-                "  \"request_id\": \"not-really-a-request-id\",\n" +
-                "  \"status\": \"test\",\n" +
-                "  \"error_text\": \"error\"\n" +
-                "}"));
-        VerifyResult r = client.verify(new VerifyRequest(
-                "447700900999",
-                "TestBrand",
-                "15555215554", 6, Locale.US));
-        assertEquals(VerifyResult.STATUS_INTERNAL_ERROR, r.getStatus());
-    }
-
-    @Test
-    public void testCheck() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(200, "<?xml version='1.0' encoding='UTF-8' ?>\n" +
-                "    <verify_response>\n" +
-                "        <event_id>verify-check-id</event_id>\n" +
-                "        <status>0</status>\n" +
-                "        <price>100.00</price>\n" +
-                "        <currency>EUR</currency>\n" +
-                "        <error_text>error</error_text>\n" +
-                "    </verify_response>"));
-
-        CheckResult c = client.check("verify-check-id", "1234", "my-ip-address");
-        assertEquals(CheckResult.STATUS_OK, c.getStatus());
-    }
-
-    @Test
-    public void testCheckMissingParams() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(200, "<?xml version='1.0' encoding='UTF-8' ?>\n" +
-                "    <verify_response>\n" +
-                "        <event_id>verify-check-id</event_id>\n" +
-                "        <status>0</status>\n" +
-                "        <price>100.00</price>\n" +
-                "        <currency>EUR</currency>\n" +
-                "        <error_text>error</error_text>\n" +
-                "    </verify_response>"));
-
-        try {
-            client.check(null, null);
-            fail("Calling check with null destination should throw IllegalArgumentException");
-        } catch (IllegalArgumentException e) {
-            // This is expected
-        }
-    }
-
-    @Test
-    public void testCheckError() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(200, "<?xml version='1.0' encoding='UTF-8' ?>\n" +
-                "    <verify_response>\n" +
-                "        <status>6</status>\n" +
-                "        <error_text>Something has gone terribly wrong!</error_text>\n" +
-                "    </verify_response>"));
-
-        CheckResult c = client.check("verify-check-id", "1234", "my-ip-address");
-        assertEquals(CheckResult.STATUS_INVALID_REQUEST, c.getStatus());
-    }
-
-    @Test
-    public void testCheckInvalidResponse() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(
-                200,
-                "<?xml version='1.0' encoding='UTF-8' ?>\n" +
-                        "<not_valid_response />"));
-
-        try {
-            client.check("verify-check-id", "1234", "my-ip-address");
-            fail("Parsing an invalid response should raise NexmoResponseParseException");
-        } catch (NexmoResponseParseException parseException) {
-            // This is expected
-        }
-    }
 
     @Test
     public void testSearchSuccess() throws Exception {
